pyproject.toml
[tool.poetry]
name = "ocr"
version = "0.1.0"
description = ""
authors = ["zuhairabbas14"]

[tool.poetry.dependencies]
python = "3.7.2"
poppler-utils = "^0.1.0"
pytorch-pretrained-bert = "^0.6.2"
pdf2image = "^1.16.0"
tesseract-ocr = "^0.0.1"
click = "^8.0.1"
pyenchant = "^3.2.1"
pytorch = "^1.0.2"
Pillow = "^8.3.1"
pytesseract = "^0.3.8"
opencv-python = "^4.5.3"
numpy = "^1.21.0"
nltk = "^3.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
split_before_logical_operator = true
